requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
joblib==1.2.0
numpy==1.25.0
packaging==23.1
pandas==2.0.2
patsy==0.5.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
